authentication page splitted, generic method for popups
