+++ automation/package.json
@@ -2,6 +2,7 @@
   "name": "automation",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {},
   "keywords": [],
   "author": "",
